:recycle: Refine DF#slice same as #remove
